Fixed problems with database module and started user api
The database object is now attached to every request under
req.app.db or if you have access to the app you can find the
database under app.db.

I also started to create the user api.
This was just a test. There still needs to be hashing implemented for
the passwords...